aider: Set up frame reading logic for the video track by initiating asset reader and writer, requesting media data, and reading video frames from the asset reader output.
# Aider chat conversation:

USER: choose the current highest priority and make an incremental improvement to it in order to replicate the original Movies.swift code in a refactored manner. make small increments to the implementation and before writing the diff, explain your intended goal of this step. at most only implement 1/20th of the final goal at each step. Before writing the diff explain the goal in englishleave a comment in the code about your intended next step, and the step after this one.

ASSISTANT: The next incremental improvement will involve setting up the frame reading logic for the video track. This step will include:

1. Initiating the asset reader and writer to start reading and writing.
2. Requesting media data when ready on the video processing queue.
3. Reading video frames from the asset reader output.
4. Preparing the next steps for handling the received video frames (which will be implemented in subsequent steps).

This step will not include the actual upscaling and writing of video frames, which will be handled later. We will also leave comments in the code indicating the next steps, which will be to upscale the video frames and append them to the asset writer input.

Here is the diff for this step:

```diff
--- FreeScaler/Movies_refactored.swift
+++ FreeScaler/Movies_refactored.swift
@@ ... @@
             assetWriter.startSession(atSourceTime: .zero)

             // TODO: Next, implement the frame reading and writing logic for video
+            // Request media data when ready on the video processing queue
+            assetWriterVideoInput.requestMediaDataWhenReady(on: self.videoQueue) {
+                // Read video frames from the asset reader output
+                while assetWriterVideoInput.isReadyForMoreMediaData {
+                    if let sampleBuffer = assetReaderVideoOutput.copyNextSampleBuffer() {
+                        // TODO: Next, upscale the video frames and append them to the asset writer input
+                        // TODO: After that, mark the asset writer input as finished when all frames are processed
+                    } else {
+                        // No more samples are available: mark the input as finished
+                        assetWriterVideoInput.markAsFinished()
+                        break
+                    }
+                }
+            }

             // TODO: After that, implement the frame reading and writing logic for audio
         }
```

This change sets up the video processing queue to request media data when ready and to read video frames from the asset reader output. It also includes comments indicating the next steps to be taken in the refactoring process.