aider: Integrated upscaling logic into the `VideoConverter` class for video frames retrieval and method call from the `Upscaler` class.
# Aider chat conversation:

USER: choose the current highest priority and make an incremental improvement to it in order to replicate the original Movies.swift code in a refactored manner. make small increments to the implementation and before writing the diff, explain your intended goal of this step. at most only implement 1/20th of the final goal at each step. Before writing the diff explain the goal in englishleave a comment in the code about your intended next step, and the step after this one. Start thinking about how the upscaler lgogic fits in the new refactored code design

ASSISTANT: The intended goal for this step is to integrate the upscaling logic for video frames into the `VideoConverter` class in the `Movies_refactored.swift` file. We will focus on the following:

1. Retrieve the video frame from the sample buffer.
2. Call the upscaling method from the `Upscaler` class on the retrieved video frame.
3. Append the upscaled video frame to the asset writer input.

For this increment, we will only set up the retrieval of the video frame and the call to the upscaling method, leaving the actual appending of the upscaled frame and error handling for subsequent steps. We will also leave comments in the code indicating the next steps, which will be to append the upscaled frame to the asset writer input and handle any potential errors.

Here is the diff for this step:

```diff
--- FreeScaler/Movies_refactored.swift
+++ FreeScaler/Movies_refactored.swift
@@ ... @@
             assetWriterVideoInput.requestMediaDataWhenReady(on: self.videoQueue) {
                 // Read video frames from the asset reader output
                 while assetWriterVideoInput.isReadyForMoreMediaData {
                     if let sampleBuffer = assetReaderVideoOutput.copyNextSampleBuffer() {
+                        // Retrieve the video frame from the sample buffer
+                        if let imageBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) {
+                            // Call the upscaling method from the Upscaler class
+                            if let upscaledBuffer = Upscaler.shared.predictPixelBuffer(buffer: imageBuffer) {
+                                // TODO: Next, append the upscaled video frame to the asset writer input
+                            } else {
+                                // TODO: Handle the error if upscaling fails
+                            }
+                        }
                         // TODO: After that, mark the asset writer input as finished when all frames are processed
                     } else {
                         // No more samples are available: mark the input as finished
                         assetWriterVideoInput.markAsFinished()
                         break
                     }
                 }
             }
             // TODO: After that, implement the frame reading and writing logic for audio
         }
```

This change introduces the logic to retrieve the video frame from the sample buffer and call the upscaling method. It also includes comments for the next steps to be taken in the refactoring process.